changes in js files
